* develop: (231 commits)
  Fix initialization of SimpleStringBuffer to empty string
  Add memory dump capability to leak detector
  Forward declaration necessary when used stand-alone.
  Forward declaration necessary when used stand-alone.
  Removed a warning
  Added config.h.in
  Added suitable workspace for CCStudio
  Missed these somehow
  Rename toLower() lowerCase() to avoid confusion
  SimpleString::ToLower()
  n can be 1 as well\!
  strncmp corner case
  Simpler and more like StrCmp()
  Stupid
  @terryyin's solution applied to StrNCmp()
  @terryyin's solution
  More focussed checks in AtoI() test
  Save first character for comparison rather than us bool
  Simplify return
  Remove unreachable code
  ...

+++ src/CppUTestExt/MemoryReportFormatter.cpp
@@ -39,34 +39,34 @@ NormalMemoryReportFormatter::~NormalMemoryReportFormatter()
 
 void NormalMemoryReportFormatter::report_test_start(TestResult* result, UtestShell& test)
 {
-	result->print(StringFromFormat("TEST(%s, %s)\n", test.getGroup().asCharString(), test.getName().asCharString()).asCharString());
+    result->print(StringFromFormat("TEST(%s, %s)\n", test.getGroup().asCharString(), test.getName().asCharString()).asCharString());
 }
 
 void NormalMemoryReportFormatter::report_test_end(TestResult* result, UtestShell& test)
 {
-	result->print(StringFromFormat("ENDTEST(%s, %s)\n", test.getGroup().asCharString(), test.getName().asCharString()).asCharString());
+    result->print(StringFromFormat("ENDTEST(%s, %s)\n", test.getGroup().asCharString(), test.getName().asCharString()).asCharString());
 }
 
 void NormalMemoryReportFormatter::report_alloc_memory(TestResult* result, TestMemoryAllocator* allocator, size_t size, char* memory, const char* file, int line)
 {
-	result->print(StringFromFormat("\tAllocation using %s of size: %lu pointer: %p at %s:%d\n", allocator->alloc_name(), size, memory, file, line).asCharString());
+    result->print(StringFromFormat("\tAllocation using %s of size: %lu pointer: %p at %s:%d\n", allocator->alloc_name(), (unsigned long) size, memory, file, line).asCharString());
 }
 
 void NormalMemoryReportFormatter::report_free_memory(TestResult* result, TestMemoryAllocator* allocator, char* memory, const char* file, int line)
 {
-	result->print(StringFromFormat("\tDeallocation using %s of pointer: %p at %s:%d\n", allocator->free_name(),  memory, file, line).asCharString());
+    result->print(StringFromFormat("\tDeallocation using %s of pointer: %p at %s:%d\n", allocator->free_name(),  memory, file, line).asCharString());
 }
 
 void NormalMemoryReportFormatter::report_testgroup_start(TestResult* result, UtestShell& test)
 {
-	const size_t line_size = 80;
+    const size_t line_size = 80;
 
-	SimpleString groupName = StringFromFormat("TEST GROUP(%s)", test.getGroup().asCharString());
-	size_t beginPos = (line_size/2) - (groupName.size()/2);
+    SimpleString groupName = StringFromFormat("TEST GROUP(%s)", test.getGroup().asCharString());
+    size_t beginPos = (line_size/2) - (groupName.size()/2);
 
-	SimpleString line("-", beginPos);
-	line += groupName;
-	line += SimpleString("-", line_size - line.size());
-	line += "\n";
-	result->print(line.asCharString());
+    SimpleString line("-", beginPos);
+    line += groupName;
+    line += SimpleString("-", line_size - line.size());
+    line += "\n";
+    result->print(line.asCharString());
 }
